Replace getters and settings in DBusRemote object and use methods instead.
Getters and setters need to be symmetrical and this is difficult and/or impossible with async behaviour.
It also can be confusing and make them not look async.
Also use 'call' as a prefix for method calls. The combination of the get/set/call prefixes is more consistent and avoids property/method name collisions.

I would recommend that developers using the generated code modify it to use properties if their API requires it, using org.freedesktop.DBus.ObjectManager or similar.