Fix DBusClient blocking when closing.
This was due to the way the signal subscription worked. When a subscription was
cancelled it would not complete until the next signal was received. This would
trigger a yield statement and close the stream.

Instead just return the stream unmodified, and don't wait to get the unique name
for the signal sender. It doesn't matter how long it takes for the unique name
to be obtained, the stream wouldn't have worked until that completed anyway.

Fixes https://github.com/canonical/dbus.dart/issues/121