Add DBusServer
This is a non-complete implementation of a basic D-Bus server, primarily here
to allow unit tests to be written. It could be improved in case an app needed
to use a private D-Bus, but it's more likely that the system/session D-Buses
will be sufficient.

Fixes https://github.com/canonical/dbus.dart/issues/12